+++ src/main/java/com/jvm_bloggers/frontend/admin_area/moderation/ModerationPageRequestHandler.java
@@ -28,7 +28,7 @@ public class ModerationPageRequestHandler implements IDataProvider<BlogPost> {
     @Override
     public Iterator<? extends BlogPost> iterator(long first, long count) {
         log.debug("Refreshing data, first {}, count {}", first, count);
-        int page = Long.valueOf(first / paginationConfiguration.getDefaultPageSize()).intValue();
+        int page = (int) (first / paginationConfiguration.getDefaultPageSize());
         long start = System.currentTimeMillis();
         Iterator<BlogPost> iterator = blogPostRepository
             .findLatestPosts(PageRequest.of(page, paginationConfiguration.getDefaultPageSize()))
